finalise conflict fixing: ready to merge

+++ to/CreateTrajectory.cpp
@@ -203,7 +203,7 @@ auto calculatedTrajectories(const std::shared_ptr<Database> &database,
                   calculated_n_1_states[0][0].item<float>(),calculated_n_1_states[0][1].item<float>()) / max(calculated_n_1_states[0][4].item<float>(),float(10))) * 100); //distance to mergeing point
     waypoint->setLongitude(ProcessedGPStoRoadUserGPS(calculated_n_1_states[0][0].item<float>()));
     waypoint->setLatitude(ProcessedGPStoRoadUserGPS(calculated_n_1_states[0][1].item<float>()));
-    waypoint->setSpeed(ProcessedSpeedtoRoadUserSpeed(max(calculated_n_1_states[0][4].item<float>(),float(10))));
+    waypoint->setSpeed(ProcessedSpeedtoRoadUserSpeed(max(calculated_n_1_states[0][4].item<float>(),float(40))));
     (!(valid_action.second)) ? waypoint->setLanePosition(mergingVehicle->getLanePosition() + 1) : waypoint->setLanePosition(mergingVehicle->getLanePosition());
 		waypoint->setHeading(ProcessedHeadingtoRoadUserHeading(calculated_n_1_states[0][19].item<float>()));
     mergingManeuver->addWaypoint(waypoint);
